trino-fix

+++ core/trino-main/src/main/java/io/trino/operator/join/unspilled/LookupJoinPageBuilder.java
@@ -40,8 +40,8 @@ public class LookupJoinPageBuilder
     private final IntArrayList probeIndexBuilder = new IntArrayList();
     private final PageBuilder buildPageBuilder;
     private final int buildOutputChannelCount;
-    private long estimatedProbeBlockBytes;
-    private long estimatedProbeRowSize = -1;
+    private int estimatedProbeBlockBytes;
+    private int estimatedProbeRowSize = -1;
     private int previousPosition = -1;
     private boolean isSequentialProbeIndices = true;
     private boolean repeatBuildRow;
@@ -242,13 +242,13 @@ public class LookupJoinPageBuilder
         }
     }
 
-    private long getEstimatedProbeRowSize(JoinProbe probe)
+    private int getEstimatedProbeRowSize(JoinProbe probe)
     {
         if (estimatedProbeRowSize != -1) {
             return estimatedProbeRowSize;
         }
 
-        long estimatedProbeRowSize = 0;
+        int estimatedProbeRowSize = 0;
         for (int index : probe.getOutputChannels()) {
             Block block = probe.getPage().getBlock(index);
             // Estimate the size of the probe row
